+++ src/translations/translations_3.rs
@@ -23,6 +23,8 @@ pub fn general_translation(language: Language) -> &'static str {
         Language::ZH => "通用",
         Language::KO => "일반",
         Language::TR => "Genel",
+        Language::PT => "Geral",
+        Language::UK => "Загальні",
         _ => "General",
     }
 }
@@ -35,6 +37,7 @@ pub fn zoom_translation(language: Language) -> &'static str {
         | Language::FR
         | Language::DE
         | Language::RO
+        | Language::PT
         | Language::SV => "Zoom",
         // Language::FA => "بزرگنمایی",
         Language::PL => "Powiększenie",
@@ -45,6 +48,7 @@ pub fn zoom_translation(language: Language) -> &'static str {
         Language::ZH => "缩放",
         Language::KO => "확대",
         Language::TR => "Yakınlaştırma",
+        Language::UK => "Масштабування",
         _ => "Zoom",
     }
 }
@@ -67,6 +71,8 @@ pub fn mmdb_files_translation(language: Language) -> &'static str {
         Language::ZH => "数据库文件",
         Language::KO => "데이터베이스 파일",
         Language::TR => "Veri tabanı dosyaları",
+        Language::PT => "Arquivos da base de dados",
+        Language::UK => "Файли бази даних",
         _ => "Database files",
     }
 }
@@ -89,6 +95,8 @@ pub fn params_not_editable_translation(language: Language) -> &'static str {
         Language::ZH => "以下参数在分析过程中不能修改",
         Language::KO => "분석 중 다음의 매개변수들은 수정할 수 없습니다",
         Language::TR => "Analiz sırasında bu parametrelere müdahale edilemez",
+        Language::PT => "Os seguintes parâmetros não podem ser modificados durante a análise",
+        Language::UK => "Наступні параметри не можна змінювати під час аналізу трафіку",
         _ => "The following parameters can't be modified during the analysis",
     }
 }
@@ -97,7 +105,7 @@ pub fn custom_style_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Custom style",
         // Language::FA => "شیوه سفارشی",
-        Language::ES => "Estilo personalizado",
+        Language::ES | Language::PT => "Estilo personalizado",
         Language::IT => "Stile personalizzato",
         Language::FR => "Style personnalisé",
         Language::DE => "Benutzerdefinierter Stil",
@@ -111,6 +119,7 @@ pub fn custom_style_translation(language: Language) -> &'static str {
         Language::ZH => "自定义样式",
         Language::KO => "사용자 지정 스타일",
         Language::TR => "Kişisel görünüm",
+        Language::UK => "Власний стиль",
         _ => "Custom style",
     }
 }
@@ -131,6 +140,8 @@ pub fn copy_translation(language: Language) -> &'static str {
         Language::ZH => "复制",
         Language::KO => "복사",
         Language::TR => "Kopyala",
+        Language::PT => "Copiar",
+        Language::UK => "Копіювати",
         _ => "Copy",
     }
 }
@@ -147,12 +158,13 @@ pub fn port_translation(language: Language) -> &'static str {
         | Language::TR => "Port",
         // Language::FA => "درگاه",
         Language::ES => "Puerto",
-        Language::IT => "Porta",
+        Language::IT | Language::PT => "Porta",
         Language::RU => "Порт",
         Language::JA => "ポート",
         Language::VI => "Cổng",
         Language::ZH => "端口",
         Language::KO => "포트",
+        Language::UK => "Порт",
         _ => "Port",
     }
 }
@@ -161,7 +173,7 @@ pub fn invalid_filters_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Invalid filters",
         // Language::FA => "صافی نامعتبر",
-        Language::ES => "Filtros inválidos",
+        Language::ES | Language::PT => "Filtros inválidos",
         Language::IT => "Filtri non validi",
         Language::FR => "Filtres invalides",
         Language::DE => "Ungültige Filter",
@@ -175,6 +187,7 @@ pub fn invalid_filters_translation(language: Language) -> &'static str {
         Language::ZH => "无效的过滤器",
         Language::KO => "잘못된 필터",
         Language::TR => "Geçersiz filtreler",
+        Language::UK => "Неправильний формат фільтрів",
         _ => "Invalid filters",
     }
 }
@@ -196,6 +209,8 @@ pub fn messages_translation(language: Language) -> &'static str {
         Language::ZH => "信息",
         Language::KO => "메시지",
         Language::TR => "Mesajlar",
+        Language::PT => "Mensagens",
+        Language::UK => "Повідомлення",
         _ => "Messages",
     }
 }
@@ -218,6 +233,8 @@ pub fn link_type_translation(language: Language) -> &'static str {
         Language::ZH => "链接类型",
         Language::KO => "링크 유형",
         Language::TR => "Link türü",
+        Language::PT => "Tipo de conexão",
+        Language::UK => "Різновид зʼєднання",
         _ => "Link type",
     }
 }
@@ -243,6 +260,8 @@ pub fn unsupported_link_type_translation<'a>(
         Language::ZH => "Sniffnet 尚不支持与此适配器关联的链接类型...",
         Language::KO => "이 어댑터와 연결된 링크 유형은 Sniffnet에서 아직 지원되지 않습니다...",
         Language::TR => "Bu adaptör ile ilişkilendirilmiş link türü henüz Sniffnet tarafından desteklenmiyor...",
+        Language::PT => "O tipo de conexão associado com este adaptador não é suportado pelo Sniffnet ainda...",
+        Language::UK => "Різновид зʼєднання, повʼязаний з даним адаптером, ще не підтримується Sniffnet-ом...",
         _ => "The link type associated with this adapter is not supported by Sniffnet yet...",
     };
 
@@ -270,6 +289,8 @@ pub fn style_from_file_translation(language: Language) -> &'static str {
         Language::ZH => "从文件中选择样式",
         Language::KO => "파일에서 스타일을 선택하세요",
         Language::TR => "Dosyadan bir görünüm seç",
+        Language::PT => "Selecionar estilo a partir de um arquivo",
+        Language::UK => "Виберіть стиль з файлу",
         _ => "Select style from a file",
     }
 }
@@ -292,6 +313,8 @@ pub fn database_from_file_translation(language: Language) -> &'static str {
         Language::ZH => "选择数据库文件",
         Language::KO => "데이터베이스 파일 선택",
         Language::TR => "Veri tabanı dosyası seç",
+        Language::PT => "Selecione um arquivo de base de dados",
+        Language::UK => "Виберіть файл бази даних",
         _ => "Select database file",
     }
 }
@@ -314,6 +337,8 @@ pub fn filter_by_host_translation(language: Language) -> &'static str {
         Language::ZH => "按网络主机筛选",
         Language::KO => "네트워크 호스트로 필터링",
         Language::TR => "Ağ sunucusuna göre filtrele",
+        Language::PT => "Filtrar por host de rede",
+        Language::UK => "Фільтр за хостом мережі",
         _ => "Filter by network host",
     }
 }
@@ -333,6 +358,8 @@ pub fn service_translation(language: Language) -> &'static str {
         Language::ZH => "服务",
         Language::KO => "서비스",
         Language::TR => "Servis",
+        Language::PT => "Serviço",
+        Language::UK => "Сервіс",
         _ => "Service",
     }
 }
@@ -354,6 +381,8 @@ pub fn export_capture_translation(language: Language) -> &'static str {
         Language::ZH => "导出捕获文件",
         Language::KO => "캡처 파일 내보내기",
         Language::TR => "Yakalanan dosyayı dışa aktar",
+        Language::PT => "Exportar arquivo capturado",
+        Language::UK => "Експорт файлу захоплення",
         _ => "Export capture file",
     }
 }
@@ -374,6 +403,8 @@ pub fn directory_translation(language: Language) -> &'static str {
         Language::ZH => "目录",
         Language::KO => "디렉토리",
         Language::TR => "Klasör",
+        Language::PT => "Diretório",
+        Language::UK => "Тека",
         _ => "Directory",
     }
 }
@@ -395,6 +426,8 @@ pub fn select_directory_translation(language: Language) -> &'static str {
         Language::ZH => "选择目标目录",
         Language::KO => "대상 디렉토리 선택",
         Language::TR => "Hedef klasörü seç",
+        Language::PT => "Selecionar diretório de destino",
+        Language::UK => "Виберіть теку призначення",
         _ => "Select destination directory",
     }
 }
@@ -416,6 +449,8 @@ pub fn file_name_translation(language: Language) -> &'static str {
         Language::ZH => "文件名",
         Language::KO => "파일 이름",
         Language::TR => "Dosya adı",
+        Language::PT => "Nome do arquivo",
+        Language::UK => "Назва файлу",
         _ => "File name",
     }
 }
@@ -437,6 +472,8 @@ pub fn thumbnail_mode_translation(language: Language) -> &'static str {
         Language::ZH => "缩略图模式",
         Language::KO => "썸네일 모드",
         Language::TR => "Küçük resim modu",
+        Language::PT => "Modo miniatura",
+        Language::UK => "Режим мініатюри",
         _ => "Thumbnail mode",
     }
 }
@@ -458,6 +495,8 @@ pub fn learn_more_translation(language: Language) -> &'static str {
         Language::ZH => "想知道更多吗？",
         Language::KO => "더 자세히 알고 싶으십니까?",
         Language::TR => "Daha fazlasını öğrenmek ister misin?",
+        Language::PT => "Quer aprender mais?",
+        Language::UK => "Бажаєте дізнатись більше?",
         _ => "Do you want to learn more?",
     }
 }