+++ src/translations/types/language.rs
@@ -126,6 +126,8 @@ impl Language {
                 | Language::ZH
                 | Language::KO
                 | Language::TR
+                | Language::PT
+                | Language::UK
         )
     }
 }